Console Color changes (description)/typo in skill multiplier fix
Changed console colors according to what is happening. System stuff will show as yellow messages, blue will be the player color and red is the enemy color.

So that way, it really feels more like a game and make it easier to the player understand what is going on. I was very confused when everything was white, I couldnt really know what was happening.